requirements.txt
google-generativeai
langchain
tiktoken
chromadb
fastapi
GitPython
python-dotenv
groq
langchain-google-genai
langchain-community